Remove warning_message alias
When running the setup script the alias warning message terminates the setup from installing as the system believes the alias for the warning_message is an actual command. 

By removing the alias warning the script will continue the setup and make the necessary changes to for the Linux system to communicate with the flipper.