Do not use a device that fails to open
When the USB device fails to open the application continues to use the
device with unfilled device data with empty serial number.

Then the application attempts to find a serial port matching the
unfilled device info structure. If you happen to have a Chinese serial
port with empty serial number the application will attempt to use it.

Invalidate the device info when filling out the data fails.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>